feat(client): Allow binding interface
