+++ examples/set_local_address.rs
@@ -1,22 +1,19 @@
 use rquest::tls::Impersonate;
+use std::net::Ipv4Addr;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Build a client to mimic Chrome126
-    #[cfg(any(target_os = "android", target_os = "fuchsia", target_os = "linux"))]
-    let client = rquest::Client::builder()
+    let mut client = rquest::Client::builder()
         .impersonate(Impersonate::Chrome126)
         .enable_ech_grease()
-        .interface("eth0")
         .permute_extensions()
         .build()?;
 
-    #[cfg(not(any(target_os = "android", target_os = "fuchsia", target_os = "linux")))]
-    let client = rquest::Client::builder()
-        .impersonate(Impersonate::Chrome126)
-        .enable_ech_grease()
-        .permute_extensions()
-        .build()?;
+    let resp = client.get("https://api.ip.sb/ip").send().await?;
+    println!("{}", resp.text().await?);
+
+    client.set_local_address(Some(Ipv4Addr::new(172, 20, 10, 2).into()));
 
     let resp = client.get("https://api.ip.sb/ip").send().await?;
     println!("{}", resp.text().await?);
